Move painel covid19 para app covid19
