+++ core/templates/home.html
@@ -38,7 +38,7 @@
             </p>
           </div>
           <div class="card-action">
-            <a href="{% url 'core:dataset-table-detail' 'covid19' 'caso' %}">Ver mais</a>
+            <a href="{% url 'covid19:dashboard' %}">Ver mais</a>
           </div>
         </div>
       </div>
