+++ core/views_special.py
@@ -7,7 +7,6 @@ from cached_property import cached_property
 from cryptography.fernet import Fernet, InvalidToken
 from django.conf import settings
 from django.contrib.postgres.search import SearchQuery
-from django.db.models import Max, Sum
 from django.http import Http404
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
@@ -274,130 +273,3 @@ def company_groups(request):
 
     context = {"form": form, "company": company, "nodes": nodes, "links": links}
     return render(request, "specials/company-groups.html", context)
-
-
-class Covid19Stats:
-
-    def __init__(self):
-        self.Boletim = get_table_model("covid19", "boletim")
-        self.Caso = get_table_model("covid19", "caso")
-
-    @cached_property
-    def city_cases(self):
-        return self.Caso.objects.filter(is_last=True, place_type="city")
-
-    @cached_property
-    def state_cases(self):
-        return self.Caso.objects.filter(is_last=True, place_type="state")
-
-    @property
-    def total_reports(self):
-        return self.Boletim.objects.count()
-
-    @property
-    def affected_cities(self):
-        return self.city_cases.count()
-
-    @property
-    def cities_with_deaths(self):
-        return self.city_cases.filter(deaths__gt=0).count()
-
-    @property
-    def state_totals(self):
-        return self.state_cases.aggregate(
-            total_confirmed=Sum("confirmed"),
-            total_deaths=Sum("deaths"),
-            total_population=Sum("estimated_population_2019"),
-        )
-
-    @property
-    def total_confirmed(self):
-        return self.state_totals["total_confirmed"]
-
-    @property
-    def total_deaths(self):
-        return self.state_totals["total_deaths"]
-
-    @property
-    def total_population(self):
-        return self.state_totals["total_population"]
-
-    @property
-    def city_totals(self):
-        return self.city_cases.aggregate(
-            max_confirmed=Max("confirmed"),
-            max_confirmed_per_100k_inhabitants=Max("confirmed_per_100k_inhabitants"),
-            max_death_rate=Max("death_rate"),
-            max_deaths=Max("deaths"),
-            total_population=Sum("estimated_population_2019")
-        )
-
-    @property
-    def affected_population(self):
-        return self.city_totals["total_population"]
-
-
-def covid19(request):
-    stats = Covid19Stats()
-    affected_cities = stats.affected_cities
-    affected_population = stats.affected_population
-    cities_with_deaths = stats.cities_with_deaths
-    total_confirmed = stats.total_confirmed
-    total_deaths = stats.total_deaths
-    total_population = stats.total_population
-
-    aggregate = [
-        {
-            "title": "Boletins coletados",
-            "value": stats.total_reports,
-            "tooltip": "Total de boletins das Secretarias Estaduais de Saúde coletados pelos voluntários",
-        },
-        {
-            "title": "Casos confirmados",
-            "value": total_confirmed,
-            "tooltip": "Total de casos confirmados",
-        },
-        {
-            "title": "Óbitos confirmados",
-            "value": total_deaths,
-            "tooltip": "Total de óbitos confirmados",
-        },
-        {
-            "title": "Municípios atingidos",
-            "value": f"{affected_cities} ({100 * affected_cities / 5570:.0f}%)",
-            "tooltip": "Total de municípios com casos confirmados",
-        },
-        {
-            "title": "População desses municípios",
-            "value": f"{affected_population / 1_000_000:.0f}M ({100 * (affected_population / total_population):.0f}%)",
-            "tooltip": "População dos municípios com casos confirmados, segundo estimativa IBGE 2019",
-        },
-        {
-            "title": "Municípios c/ óbitos",
-            "value": f"{cities_with_deaths} ({100 * cities_with_deaths / affected_cities:.0f}%)",
-            "tooltip": "Total de municípios com óbitos confirmados (o percentual é em relação ao total de municípios com casos confirmados)",
-        },
-    ]
-
-    value_keys = ("confirmed", "deaths", "death_rate", "confirmed_per_100k_inhabitants")
-    city_values = {key: {} for key in value_keys}
-    city_data = []
-    # TODO: what should we do about "Importados/Indefinidos"?
-    for row in stats.city_cases.filter(city_ibge_code__isnull=False):
-        row = row.__dict__
-        for key in value_keys:
-            row[key] = row[key] or 0
-            city_values[key][row["city_ibge_code"]] = row[key]
-        row["death_rate_percent"] = row.pop("death_rate") * 100
-        row["date_str"] = str(row["date"])
-        row["city_str"] = slugify(row["city"])
-        year, month, day = row["date_str"].split("-")
-        row["date"] = datetime.date(int(year), int(month), int(day))
-        city_data.append(row)
-    max_values = {key: max(city_values[key].values()) for key in value_keys}
-
-    return render(
-        request,
-        "specials/covid19.html",
-        {"city_data": city_data, "aggregate": aggregate, "city_values_json": json.dumps(city_values), "max_values_json": json.dumps(max_values)},
-    )
