+++ covid19/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import views
+
+
+app_name = "covid19"
+urlpatterns = [
+    path("", views.dashboard, name="dashboard"),
+]
